% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1.R
\name{add1}
\alias{add1}
\title{addition function}
\usage{
add1(x, y)
}
\arguments{
\item{x}{The first value to add.}

\item{y}{The second value to add}
}
\value{
A single value that is the sum of the 2 numbers
}
\description{
This function adds together 2 numbers
}
