% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply1}
\alias{multiply1}
\title{Multiply 2 numbers}
\usage{
multiply1(x, y)
}
\arguments{
\item{x}{The first number to multiply}

\item{y}{The second number to multiply}
}
\value{
THe multiplication of the 2 numbers
}
\description{
This function multiplies 2 numbers together
}
