% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1.R
\name{add_3_nums}
\alias{add_3_nums}
\title{Add 3 numbers}
\usage{
add_3_nums(x, y, z)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}

\item{z}{The third number}
}
\value{
A single value that is the sum of the 3 numbers
}
\description{
Function adds together 2 numbers
}
\author{
Kathryn
}
